Move sims from six_hump_camel to var_resources
