+++ libensemble/sim_funcs/six_hump_camel.py
@@ -8,32 +8,21 @@ Six-hump camel function is documented here:
 __all__ = [
     "six_hump_camel",
     "six_hump_camel_simple",
-    "six_hump_camel_with_variable_resources",
-    "six_hump_camel_GPU_variable_resources",
-    "six_hump_camel_CUDA_variable_resources",
     "persistent_six_hump_camel",
 ]
 
-import os
 import sys
 import time
 import numpy as np
 
-from libensemble.executors.executor import Executor
 from libensemble.message_numbers import (
     EVAL_SIM_TAG,
     FINISHED_PERSISTENT_SIM_TAG,
     PERSIS_STOP,
     STOP_TAG,
-    TASK_FAILED,
-    UNSET_TAG,
-    WORKER_DONE,
 )
-from libensemble.resources.resources import Resources
 from libensemble.tools.persistent_support import PersistentSupport
-from libensemble.tools.test_support import check_gpu_setting
 
-#TODO May remove/move six_hump_camel_with_variable_resources / six_hump_camel_CUDA_variable_resources
 
 def six_hump_camel(H, persis_info, sim_specs, libE_info):
     """
@@ -77,177 +66,6 @@ def six_hump_camel_simple(x, _, sim_specs):
         time.sleep(sim_specs["user"]["pause_time"])
 
     return H_o
-
-
-def six_hump_camel_with_variable_resources(H, _, sim_specs):
-    """
-    Evaluates the six hump camel for a collection of points given in ``H["x"]``
-    via the executor, supporting variable sized simulations/resources, as
-    determined by the generator.
-
-    .. seealso::
-        `test_uniform_sampling_with_variable_resources.py <https://github.com/Libensemble/libensemble/blob/develop/libensemble/tests/regression_tests/test_uniform_sampling_with_variable_resources.py>`_ # noqa
-    """
-
-    batch = len(H["x"])
-    H_o = np.zeros(batch, dtype=sim_specs["out"])
-    app = sim_specs["user"].get("app", "helloworld")
-    dry_run = sim_specs["user"].get("dry_run", False)  # dry_run only prints run lines in ensemble.log
-    set_cores_by_rsets = True  # If True use rset count to set num procs, else use all available to this worker.
-    core_multiplier = 1  # Only used with set_cores_by_rsets as a multiplier.
-
-    exctr = Executor.executor  # Get Executor
-    task_states = []
-    for i, x in enumerate(H["x"]):
-        nprocs = None  # Will be as if argument is not present
-        if set_cores_by_rsets:
-            resources = Resources.resources.worker_resources
-            nprocs = resources.num_rsets * core_multiplier
-
-        inpt = None  # Will be as if argument is not present
-        if app == "six_hump_camel":
-            inpt = " ".join(map(str, H["x"][i]))
-
-        task = exctr.submit(
-            app_name=app,
-            app_args=inpt,
-            num_procs=nprocs,
-            stdout="out.txt",
-            stderr="err.txt",
-            dry_run=dry_run,
-        )
-
-        if not dry_run:
-            task.wait()  # Wait for run to complete
-
-            # while(not task.finished):
-            #     time.sleep(0.1)
-            #     task.poll()
-
-        task_states.append(task.state)
-
-        if app == "six_hump_camel":
-            # H_o["f"][i] = float(task.read_stdout())  # Reads whole file
-            with open("out.txt") as f:
-                H_o["f"][i] = float(f.readline().strip())  # Read just first line
-        else:
-            # To return something in test
-            H_o["f"][i] = six_hump_camel_func(x)
-
-    calc_status = UNSET_TAG  # Returns to worker
-    if all(t == "FINISHED" for t in task_states):
-        calc_status = WORKER_DONE
-    elif any(t == "FAILED" for t in task_states):
-        calc_status = TASK_FAILED
-
-    return H_o, calc_status
-
-def six_hump_camel_GPU_variable_resources(H, persis_info, sim_specs, libE_info):
-    """Launches an app and automatically assigns GPU resources.
-
-    The six_hump_camel app does not run on the GPU, but this test demonstrates
-    how to automatically assign the GPUs given to this worker via the MPIExecutor.
-    The method used to assign GPUs will be determined by the MPI runner or other
-    detected system parameters. It may also be explicitly configured (e.g., by
-    setting the LIBE_PLATFORM environment variable).
-
-    See the forces_gpu tutorial for an example with an actual GPU application.
-    See six_hump_camel_CUDA_variable_resources for an example where the sim function
-    interrogates available resources and sets explicitly.
-    """
-    x = H["x"][0]
-    H_o = np.zeros(1, dtype=sim_specs["out"])
-    dry_run = sim_specs["user"].get("dry_run", False)  # logs run lines instead of running
-    inpt = " ".join(map(str, x)) # Application input
-
-    exctr = Executor.executor  # Get Executor
-
-    # Launch application via system MPI runner, using assigned resources.
-    task = exctr.submit(
-        app_name="six_hump_camel",
-        app_args=inpt,
-        auto_assign_gpus = True,
-        match_procs_to_gpus = True,
-        stdout="out.txt",
-        stderr="err.txt",
-        dry_run = dry_run,
-    )
-
-    if not dry_run:
-        task.wait()  # Wait for run to complete
-
-        # Access app output
-        with open("out.txt") as f:
-            H_o["f"] = float(f.readline().strip())  # Read just first line
-
-    # Asserts GPU set correctly (for default MPI settings)
-    check_gpu_setting(task, print_setting=True)
-
-    calc_status = WORKER_DONE if task.state == "FINISHED" else "FAILED"
-    return H_o, persis_info, calc_status
-
-
-def six_hump_camel_CUDA_variable_resources(H, _, sim_specs, libE_info):
-    """Launches an app setting GPU resources
-
-    The standard test apps do not run on GPU, but demonstrates accessing resource
-    information to set ``CUDA_VISIBLE_DEVICES``, and typical run configuration.
-
-    For an equivalent function that auto-assigns GPUs using platform detection, see
-    six_hump_camel_GPU_variable_resources.
-    """
-    x = H["x"][0]
-    H_o = np.zeros(1, dtype=sim_specs["out"])
-    dry_run = sim_specs["user"].get("dry_run", False)  # dry_run only prints run lines in ensemble.log
-
-    # Interrogate resources available to this worker
-    resources = Resources.resources.worker_resources
-    slots = resources.slots
-
-    assert resources.matching_slots, f"Error: Cannot set CUDA_VISIBLE_DEVICES when unmatching slots on nodes {slots}"
-
-    num_nodes = resources.local_node_count
-
-    # Set to slots
-    resources.set_env_to_slots("CUDA_VISIBLE_DEVICES")
-    cores_per_node = resources.slot_count
-
-    # Set to detected GPUs
-    # gpus_per_slot = resources.gpus_per_rset
-    # resources.set_env_to_slots("CUDA_VISIBLE_DEVICES", multiplier=gpus_per_slot)
-    # cores_per_node = resources.slot_count * gpus_per_slot  # One CPU per GPU
-
-    print(
-        f"Worker {libE_info['workerID']}: CUDA_VISIBLE_DEVICES={os.environ['CUDA_VISIBLE_DEVICES']}"
-        f"\tnodes {num_nodes} ppn {cores_per_node}  slots {slots}"
-    )
-
-    # Create application input file
-    inpt = " ".join(map(str, x))
-    exctr = Executor.executor  # Get Executor
-
-    # Launch application via system MPI runner, using assigned resources.
-    task = exctr.submit(
-        app_name="six_hump_camel",
-        app_args=inpt,
-        num_nodes=num_nodes,
-        procs_per_node=cores_per_node,
-        stdout="out.txt",
-        stderr="err.txt",
-        dry_run = dry_run,
-        # extra_args='--gpus-per-task=1'
-    )
-
-    if not dry_run:
-        task.wait()  # Wait for run to complete
-
-        # Access app output
-        with open("out.txt") as f:
-            H_o["f"] = float(f.readline().strip())  # Read just first line
-
-    calc_status = WORKER_DONE if task.state == "FINISHED" else "FAILED"
-    return H_o, calc_status
-
 
 
 def persistent_six_hump_camel(H, persis_info, sim_specs, libE_info):
