+++ libensemble/tests/functionality_tests/test_mpi_gpu_settings.py
@@ -42,7 +42,7 @@ import numpy as np
 # Import libEnsemble items for this test
 from libensemble.libE import libE
 from libensemble.sim_funcs import six_hump_camel
-from libensemble.sim_funcs.six_hump_camel import six_hump_camel_GPU_variable_resources as sim_f
+from libensemble.sim_funcs.var_resources import six_hump_camel_GPU_variable_resources as sim_f
 from libensemble.gen_funcs.persistent_sampling import uniform_random_sample_with_variable_resources as gen_f
 from libensemble.alloc_funcs.start_only_persistent import only_persistent_gens as alloc_f
 from libensemble.tools import parse_args, save_libE_output, add_unique_random_streams
